requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
flask-sqlalchemy==3.0.5
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing-extensions==4.7.1
validators==0.21.2
werkzeug==2.3.7
zipp==3.16.2
